Renaming packages (please squash all the committs before merge)
