+++ transport-native-epoll/src/main/java/io/netty/jni/libaio/ErrorInfo.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.channel.libaio;
+package io.netty.jni.libaio;
 
 /**
  * In case of a failure the callback will be replaced by an instance of this class.
