+++ transport-native-epoll/src/main/java/io/netty/jni/libaio/LibaioContext.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.channel.libaio;
+package io.netty.jni.libaio;
 
 import io.netty.channel.epoll.Native;
